chore(cleanup): refactor database connection and schema imports to use chore module
