+++ src/chore/database/schema.py
@@ -2,7 +2,7 @@ from pathlib import Path
 import sqlite3
 import logging
 from utils.logging.logging_tools import log_fn
-from utils.data.database_connection import DatabaseConnection
+from chore.database.connection import DatabaseConnection
 import config
 
 logger = logging.getLogger(__name__)
