+++ src/utils/data/database/counterparty_utils.py
@@ -2,7 +2,7 @@ import sqlite3
 from pathlib import Path
 import logging
 from typing import List, Tuple, Union, Optional
-from utils.data.database_connection import DatabaseConnection
+from chore.database.connection import DatabaseConnection
 import config
