+++ src/utils/data/database/transaction_utils.py
@@ -1,7 +1,7 @@
 import sqlite3
 from pathlib import Path
 import logging
-from utils.data.database_connection import DatabaseConnection
+from chore.database.connection import DatabaseConnection
 import config
